made landing page better
